Add RevPi SOS report plugins
The RevPi plugins for sos report are useful tools for supporting customers
during troubleshooting and often used by our support agents. In the past these
plugins were part of the revpi-tools package. In order to strip revpi-tools to
a bare minimum (for example for OEM customers), these plugins are moved into
a separate package called revpi-sos-report.

Starting with Debian Bullseye the plugin structure of sos report has changed.
This is fixed by a rewrite of the old plugins. While doing this the monolithic
plugin is broken into three smaler plugins with a specific speciality
(eg. CODESYS or HAT EEP).

This is the initial commit of the new repository.

Signed-off-by: Nicolai Buchwitz <n.buchwitz@kunbus.com>