doc: rewrite revpi-sos man page
The man page is most likely read by users that need support for their
device, so write the man page for them directly. This means it should
mention when an SOS report should be created and where it should be sent
to.
Because it can't be assumed that the user is intimately familiar with
"tar" also give examples for viewing the content and extracting the
resulting archive with it.

Signed-off-by: Thomas Böhler <t.boehler@kunbus.com>