chore: avoid unnecessary information from sos
The sos command outputs a bunch of information that we don't output to
stdout and only filter for the location of the generated sos report.
Avoid outputting all the other information that we don't need by passing
the "-q" flag.

This is a sample output with the "-q" flag included:

    pi@RevPi106245:/var/tmp$ sudo sos report -qz gzip -o revpi,revpi-eep,revpi-codesys --batch

    sosreport (version 4.0)

    Your sosreport has been generated and saved in:
            /var/tmp/sosreport-RevPi106245-2025-03-21-hbngrrb.tar.gz

Signed-off-by: Thomas Böhler <t.boehler@kunbus.com>