fix: disable size limit of logs
Logs should *always* be collected in full and not trimmed by sosreport.
This is not currently the case, and logs that are too long are trimmed
by sosreport.

This is not what we want. Disable the log limit completely to make sure
logs are always collected with *all* of their content.

Signed-off-by: Thomas Böhler <t.boehler@kunbus.com>
(cherry picked from commit 63c01a4bb8315376da329d18363d16a265102b30)